+++ src/main/java/kitchenpos/shared/external/infra/ProfanityChecker.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.external.infra;
+package kitchenpos.shared.external.infra;
 
 public interface ProfanityChecker {
     boolean containsProfanity(String text);
