+++ src/test/java/kitchenpos/products/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.application;
 
-import kitchenpos.common.external.infra.ProfanityChecker;
+import kitchenpos.shared.external.infra.ProfanityChecker;
 
 import java.util.Arrays;
 import java.util.List;
